+++ app/pages/Accounts/AccountBalanceView/AccountBalanceView.stories.tsx
@@ -76,6 +76,7 @@ const accountInfo: AccountInfo = {
     accountEncryptedAmount: {} as any,
     accountCooldowns: [],
     accountAvailableBalance: CcdAmount.fromCcd(1000),
+    accountTokens: [],
 } as AccountInfo;
 
 export const SingleSig = Template.bind({});