+++ app/preload/ledger/ledger.ts
@@ -1,4 +1,5 @@
 import { ValidatorScoreParameters } from '@concordium/web-sdk';
+
 import { Buffer } from 'buffer/';
 import EventEmitter from 'events';
 import {
@@ -310,6 +311,17 @@ export default function exposedMethods(
             keypath: number[]
         ) => {
             return getLedgerClient().signValidatorScoreParameters(
+                transaction,
+                serializedPayload,
+                keypath
+            );
+        },
+        signCreatePlt: (
+            transaction: string,
+            serializedPayload: Buffer,
+            keypath: number[]
+        ) => {
+            return getLedgerClient().signCreatePlt(
                 transaction,
                 serializedPayload,
                 keypath